test: fix failing nightly test splits
